LeetCode[clean]

+++ src/main/java/com/kristurek/leetcode/Solution2.java
@@ -52,8 +52,8 @@ public class Solution2 {
 		if (x < 0)
 			return false;
 
-		int reverse = 0;
-		int number = x;
+		long reverse = 0;
+		long number = x;
 
 		while (x > 0) {
 			reverse = reverse * 10 + x % 10;
